refactor : redis prefix 패키지 이동 및 bootcamp key 추가
